todo_service/requirements.txt
Django==3.2
Markdown==3.3.6
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0

npx==10.2.2

